Manually disable sms auth method
SMS is no longer considered a secure 2FA method, as it is too easy
for malicious actors to take over the user's cellular phone.  The
cellular providers are complicit in that they enable this takeover
with a minimal amount of information about the legitimate user.
As a result of this, an attacker can steal a user's identity,
switch the user's cell service over to a device they control, and
essentially intercept all of that user's SMS messages.  This is not
a theoretical attack and has been performed within 15 minutes as
users watch, unable to stop it.

While Duo MFA provides the ability to disable specific authentication
methods, the free level disables this ability.  Best practice is to
provide a default level that is secure, and not insist on a paid
service level in order for a user to be able to secure their service.

This patch simply disables the use of sms as a user's password.  It
is not recommended as a merge candidate for the master branch, but
as an option for users who wish to disable SMS on their free Duo
service level.  Before merging to master, it may be wise to provide
a mechanism for optionally disabling it via configuration, rather
than requiring a recompile.